+++ frontend/app/artikel/[slug]/page.module.css
@@ -50,30 +50,6 @@
     }
 }
 
-.youtubeSection {
-    margin-top: 2rem;
-    max-width: 800px;
-    display: flex;
-    flex-direction: column;
-    gap: 2rem;
-}
-
-.youtubeSection > div {
-    position: relative;
-    width: 100%;
-    padding-bottom: 56.25%; /* 16:9 aspect ratio */
-    height: 0;
-    overflow: hidden;
-}
-
-.youtubeSection > div > iframe {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-}
-
 @media (max-width: 640px) {
     .article {
         padding: 0 1rem;
@@ -94,12 +70,6 @@
 
     .coverImage {
         border-radius: 0;
-    }
-
-
-    .youtubeSection {
-        margin-top: 1.5rem;
-        gap: 1.5rem;
     }
 }
