+++ frontend/app/podcasts/[slug]/page.module.css
@@ -79,7 +79,6 @@
         border-radius: 0;
     }
 
-
     .player {
         border-radius: 0;
         border-left: none;
